requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
schedule>=1.2.0
tqdm>=4.66.0
nltk>=3.8.0
textblob>=0.17.0
vaderSentiment>=3.3.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
pytest>=7.4.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
